fix devsite
